[test_runner] Split pipeline and router in different files
Pipeline and routing logic are now split into different files.
Unit tests for the routing logic have been added.
